adds gitignore and ignores venv and mypy
Since virtual environments are not compatible between Windows and Os X, the
gitignore has been set up to ignore `env/`. The README will be updated with
instructions on how to set up the virtual environment per operating system.
Each developer will have to install the dependencies with `pip` after starting
the virtual environment.

Also, mypy will be used for testing static types in some of the games and in the
UI. Since mypy generates .mypy_cache, which is not relevant to the project,
`.mypy_cache/` has been added to the gitignore, as well.