splits chess from develop
before a simple test game was created for develop branch, chess was used as the
test. This splits chess from develop into its own feature/ branch.

It contains a basic working chess game with a base board and no move validation
or chess engine yet. A refactor is under way to make Color and PieceType enums
play nicer as well as to make Piece and BaseBoard more easier to work with.

Also, a Board and maybe a Game class will be developed soon to make it more
complete by providing move validation and status.