Makes position relative for game
When a game's `onevent` is called, if the event is a MOUSE event, then the
position is adjusted to be a realtive position with respect to the top-left
corner of the game. An additional attribute , `abs_pos`, is added in case the
original position is desired.

testgame demonstrates this new bug fix by allowing the user to left-click and
drag the square around the screen.