grow UI if game large
In addition to some cleanup, the UI now expands if the game is large enough.
Since resizing the display causes the background to be black, a `self.refill` flag
is also used to keep track of whether entire background needs to be re-filled.

Now the testgame is much larger to demonstrate what happens when a larger game
is registered with the API.

Cleanup:
- pygame.display no longer tries to use HWSURFACE (hard-ware surface) since we're probably not gaining anything.
- pygame.display no longer uses DOUBLEBUF for main surface since that is only applicable if using HWSURFACE.
- print statement removed from plethoraAPI.PlethoraAPI.import_game().
- local gameflip flag removed from onrender since unused.