refactor chess.py to be more concise
- Square:

Now each `Square` is an `enum.IntFlag` with an `index` attribute if it is one of the board `Square`s
(*A1* through *H8*). `Square`s can be composed to create `BitBoard`s, since they are
`enum.IntFlags`, and `BaseBoard` now uses this method exclusively, and the integrity of this
convention should be respected for clarity.

- Color, PieceType, Piece:

The `Color`is almost the same as before. `Piece` is now `PieceType` as an `enum.Enum`, which
represents a piece sans color, while `Piece` is an class that just uniquely identifies a piece since
it has a `Color` and a `PieceType`. `Piece` also defines `__slots__`, which could potentially result
in a significant speed up should millions of `Piece`s be created. Since `Color` and `PieceType` are
enums (using the standard library), they should be fairly efficient, as well.

- __getitem__, __setitem__, __iter__:

`BaseBoard` now supports iterating over all of the squares and converting itself to a list by
defining the `__iter__` function. For example, assuming `b = BaseBoard()`, `print(list(b)` prints a
list of all of the pieces from the top-left (`Square.A8`) to the bottom-right (`Square.H1`). You can
also use, `for p in b: print(p)` to iterate over all of the boards.

One can now access `BaseBoard` directly to get or set an item. For example, `b = Baseboard();
print(b[Square.A8])` will access the top-left square, *A8* `Square`. Similarly, `b[Square.E4] =
Piece(Color.LIGHT, PieceType.PAWN)` will set the *E4* `Square` to a light pawn. `__getitem__` and
`__setitem__` directly call `get_piece_at` and `set_piece_at` respectively and are only provided for
convenience.

- get_san, set_san

`BaseBoard`'s member functions, `get_san` and `set_san`, have remained relatively unchanged.
Although, `get_san` is ever so slightly faster. However, it should be noted that both `get_san` and
`set_san` are very slow and shouldn't be used by the engine repeatedly. Instead, the engine should
favor `get_bitboards` and `set_bitboards`, since they are much faster by a magnitude of roughly
one-hundred. See next bullet for more information on `get_bitboards` and `set_bitboards`.

- get_bitboards, set_bitboards

As mentioned above, there is now a much faster alternative to `get_san` and `set_san`. While
`get_san` and `set_san` are convenient for humans and readability, `get_bitboards` and
`set_bitboards` are much quicker. Respectively, they use `struct.pack` to pack all of the bitboards
into one sequence of bytes and use `struct.unpack` to unpack all of the bitboards into the
`BaseBoard`'s bitboards. Note that it does convert each unpacked number in the unpack (in
`set_bitboards`) to a `Square` to maintain integrity in that regard.

- pretty

`BaseBoard`'s `__repr__` is now just `"BaseBoard({!r})".format(self.get_san())` since this is more
logical. Now, `BaseBoard#pretty` should be used to print out all of the `BitBoard`s. `pretty` does
not print the `BitBoard`s as hex numbers but just the 8x8 representation of 1s and 0s. Note that
this occupies a lot of horizontal space and is quite slow; therefore it should really only be used
for debugging.

- get_*_at, set_*_at,

`BaseBoard` now has `get_color_at`, `get_piecetype_at`, and `get_piece_at`, which return the
`Color`, `PieceType`, and `Piece` at a `Square`. These convenience methods discourage direct access
of `BaseBoard` bitboards and are as fast as I could make them since they will be called very
frequently.

`set_piece_at` is also capable of setting a `Piece` at a `Square`.

As mentioned, `get_piece_at` and `set_piece_at` are abstracted one level further to `__getitem__`
and `__setitem__` for convenience.

- move

`BaseBoard#move` is a very simple method and provides no checks to verify that the piece you want to
move from one `Square` to another is a valid move. In fact, it's so rudimentary that (presume `b =
BaseBoard()`) `b.move(Square.E2, Square.E4)` is the same thing as `b[E4] = b[E2]; b[E2] = None`.

The reason `move` is so rudimentary is to make it fast. A chess editor might want to call this
directly, and in the future the chess game class (as of yet, undefined) will provide all move
validation.

- TODO
    - marry new BaseBoard with plethoraAPI using old main.py as reference
    - add doc strings again (sorry: just been hacking away and being a perfectionist - no time for docstrings :[)
    - when plethoraAPU supports: use `dirtyRects` to only redraw rects that need it
    - create chess game wrapper, if you will, around BaseBoard for move validation and the like
    - build chess engine