adds UI functionality for chess
- refactor

First, main.py is refactored to game.py, which now appropriately integrates with the plethora API.
The UI code in game.py is pretty much completely refactored and changed. More variables have been
added at the module level for variables that do not need to be modified; all of these module
variables are now upper-case with underscores to denote that they are module-variables.

- chess UI

The chess UI is much more robust. As before, the user can navigate the board by moving a
cursor using the arrow keys. The cursor is now dark green, but the concept is the same.

However, mouse events are now supported. Mouse drag and click events are handled slightly
differently and can be used in conjunction with keyup and keydown events. If the user moves or
attempts to move a piece, then the cursor is hidden until a keyup or keydown is issued. This looks
cleaner.

When the user presses the mouse down and drags a piece, a ghost piece is created at the original
square, and the user can drag the piece around. When the user lets go, various cases are handled. If
the user presses the mouse down on a piece and drags the piece to a different square and lets go
then the piece is moved to that square. If the user presses the mouse down, moves to a different
square, moves back to the original square, and finally lets go then nothing is selected. If the user
presses the mouse down, does *not* move to a different square, and lets go of the mouse button then
the piece is selected but not moved.

Lastly, all mouse events account for when the events are outside of the game window or squares on the board
(since there is a small border around the edge of the board).

- images

I don't know why the images/ directory was not added before, but it is added now. Also, all of the
border images are no longer used so they are not in the directory anymore.