API traceback and mouseup hiding
When a game fails to load, plethoraAPI now uses the traceback module to provide more useful
debugging information.

Also, when the user clicks a button, the MOUSEDOWN is listened for, but this means that since the
game is started on MOUSEDOWN, then when the user releases the mouse, a MOUSEUP event is sent
erroneously to the game. Therefore, the first MOUSEUP is ignored.