Mostly complete move validation
- Move validation

With the exception of checking for check, discovered check, and check mate, move validation is
complete and rather robust. The move validation in this commit is concerned with basic piece
movement, piece collision (only knights can "jump over" other pieces), and a few special cases
excluding any sort of check.

Excluding check, discrovered check, and check mate, basic piece movement means the following: no
piece may be moved onto a piece of the same color; only the appropriate color may move (light may
not move on dark's turn and vice versa); light pawns only advance northward; black pawns only
advance southward; pawns may not capture by advancing along a file; a pawn may always advance N or S
(for light or dark resp.) unless, as mentioned, any piece is on the target square (even an enemy
piece), since pawns cannot capture vertically but rather diagonally; a pawn may advance two squares
on the first move unless there is a piece on the skipped square or the target square, since pawns
cannot capture vertically; a pawn can capture an opponent's piece on NW|NE or SW|SE (for light or
dark resp.); a knight may move to NNW|NWW|NNE|NEE|SSW|SWW|SSE|SEE; a bishop may move along either
diagonal unless blocked by any piece but may capture an enemy piece on the target square; a rook may
move along the rank or file unless blocked by any piece but may capture an enemy piece on the target
square; a queen may move along either diagonal, the rank, or the file unless blocked by any piece
but may capture an enemy piece on the target square; a king may move to NW|N|NE|W|E|SW|S|SE.

Additionally, all other special conditions except check, discovered check, and check mate are
handled. When a pawn advances two squares on the first move, the en passant square is set. If an
enemy pawn is adjacent to the pushed pawn, then it may capture the pushed pawn on the next move via
the en passant rule. Another special case that is handled is castling. When a king is moved from the
E file to the G or C file then the king may "castle" king- or queen-side (resp.). When a king
castles, the king is moved to the target square (which is two squares E or W for king- or queen-side
castling), then the rook is placed on the "inside" of the king AS LONG AS neither the king nor the
rook with which the king is "castling" have moved. So, if the king is moved to the G file, then the
rook is moved to the F file, and if the king is moved to the C file, then the rook is moved to the
D file.

- Square

Square has received a major update to make move validation ahead of time. Each Square now has all of
the following members, which I am calling the extended cardinal set: N, W, S, E, NW, NE, SW, SE,
NNW, NWW, NNE, NEE, SSW, SWW, SSE, SEE, NN, SS. This makes it easy to access the square that is at
the given location. Note that if the cardinal, intercardinal, or extended cardinal direction would
not be defined then Square(0) is assigned, which is which is Falsy and therefore the Square
equivalent to None.

In addition to the extended cardinal directions above, each Square also has a few BitBoards to make
move validation easier. These BitBoards are denoted with a "BB_" prefix and include the following
bullted items:

+ BB_KING: all Squares a to which a king may move (excluding castling)
+ BB_KNIGHT: all Squares to which a knight may move
+ BB_FILE: all Squares in the file
+ BB_RANK: all Squares in the rank
+ BB_FDIAG: all Squares in the falling diagonal (starts at NW-most; ends at SE-most)
+ BB_RDIAG: all Squares in the rising diagonal (starts at SW-most; ends at NE-most)
+ BB_ORTHOGONALS: BB_FILE and BB_RANK
+ BB_DIAGS: BB_FDIAG and BB_RDIAG
+ BB_LINES: BB_ORTHOGONALS and BB_DIAGS
+ BB_ALL: BB_LINES and BB_KNIGHT

- pretty printing

For debugging, Square#pprint() and BaseBoard#pprint() functions have been implemented. Square has
some additional functions like pretty_list() which returns a list of the characters so that this can
be more used more easily programatically. Also, BaseBoard has pretty_bitboards() still and also
pprint_bitboards() which is similar to pprint() but it prints all of the BitBoards instead of the
characters of the pieces on the board.

- Todo

Perhaps I bust out valid_move() into valid_pseudomove() and valid_move() in order to gain speed in
engine.

Unimplemented todo from previous commit:

Need to overwrite some BaseBoard functions to prevent directly setting pieces in the Board unless it
is editable. During a game, Board#editable should be set to False, which helps maintain state.

Improve move validation in Board#valid_move() using chessprogramming.org.

Add more functions like get_algebraic_move(_:Square, _:Square) which will return the algebraic
representation of the move. This could, for example, be called before super(BaseBoard, self).move()
but after move validation with Board#valid_move().. Probably implement this in the Move() class,
however.

Maybe add Move class to store a move. (Consider using __slots__.)

Keep utilizing chessprogramming.org to make engine.